refactor(ks-vm): refactoring vm to work with new vm

+++ workspace/ks-core/src/compiler_new/environment.rs
@@ -34,7 +34,7 @@ impl Environment {
     pub fn current(&self) -> KsResult<VariableId> {
         let variables = self.last_function()?;
 
-        Ok(self.function_variables_len(variables))
+        Ok(self.function_variables_len(variables) as u64)
     }
 
     pub fn functions(self) -> HashMap<String, Pointer> {
